+++ packages/logd_sqlite/test/sqlite_sink_test.dart
@@ -2,25 +2,12 @@
 // Use of this source code is governed by a BSD-3-Clause license that can be
 // found in the LICENSE file.
 
-import 'dart:ffi';
-
 import 'package:logd/logd.dart';
 import 'package:logd_sqlite/logd_sqlite.dart';
-import 'package:sqlite3/open.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
 
 void main() {
-  setUpAll(() {
-    open.overrideFor(OperatingSystem.linux, () {
-      try {
-        return DynamicLibrary.open('libsqlite3.so');
-      } catch (_) {
-        return DynamicLibrary.open('libsqlite3.so.0');
-      }
-    });
-  });
-
   group('SqliteSink Production Suite', () {
     late Database memoryDb;
     late LogPipelineFactory factory;
@@ -32,7 +19,7 @@ void main() {
 
     tearDown(() {
       try {
-        memoryDb.dispose();
+        memoryDb.close();
       } catch (_) {}
     });
 
@@ -181,12 +168,10 @@ void main() {
       // Test distinct logger names helper
       final distinctLoggers = sink.fetchDistinctLoggerNames();
       expect(distinctLoggers.length, equals(2));
-      final names =
-          distinctLoggers.map((final l) => l['logger_name']).toList();
+      final names = distinctLoggers.map((final l) => l['logger_name']).toList();
 
       expect(names, containsAll(['auth.service', 'db.service']));
     });
-
 
     test('should apply maxEntries retention policy on flush', () async {
       final sink = SqliteSink(